% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_projected.R
\name{geo_buffer}
\alias{geo_buffer}
\title{Perform a buffer operation on a temporary projected CRS}
\usage{
geo_buffer(shp, dist = NULL, ...)
}
\arguments{
\item{shp}{A spatial object with a geographic CRS (e.g. WGS84)
around which a buffer should be drawn}

\item{dist}{The distance (in metres) of the buffer (when buffering simple features)}

\item{...}{Arguments passed to the buffer (see `?rgeos::gBuffer` or `?sf::st_buffer` for details)}
}
\description{
This function solves the problem that buffers will not be circular when used on
non-projected data.
}
\details{
Requires recent version of PROJ (>= 6.3.0).
Buffers on `sf` objects with geographic (lon/lat) coordinates can also
be done with the [`s2`](https://r-spatial.github.io/s2/) package.
}
\examples{
lib_versions <- sf::sf_extSoftVersion()
lib_versions
nc = sf::st_read(system.file("shape/nc.shp", package="sf"))
if (lib_versions[3] >= "6.3.1") {
  buff_sf <- geo_buffer(nc, dist = 50)
  plot(buff_sf$geometry)
  geo_buffer(nc$geometry, dist = 50)
  # on legacy sp objects (not tested)
  # buff_sp <- geo_buffer(routes_fast, width = 100)
  # class(buff_sp)
  # plot(buff_sp, col = "red")
}
}
