\name{sfextras-package}
\alias{sfextras-package}
\alias{sfextras}
\docType{package}
\title{\packageTitle{sfextras}}
\description{\packageDescription{sfextras}}
\section{Package Content}{\packageIndices{sfextras}}
\author{\packageAuthor{sfextras}}
\section{Maintainer}{\packageMaintainer{sfextras}}
\keyword{package}
