% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{st_centroid_coords}
\alias{st_centroid_coords}
\title{Extract centroid coordinates as a matrix of an sf polygon object for plotting a connectivity graph}
\usage{
st_centroid_coords(x)
}
\arguments{
\item{x}{sf object}
}
\value{
matrix of lon/lat coordinates
}
\description{
Extract centroid coordinates as a matrix of an sf polygon object for plotting a connectivity graph
}
\examples{
coords <- st_centroid_coords(ncovr)
ncovr_rook <- st_rook(ncovr)
rook_nb <- st_as_nb(ncovr_rook)

plot(rook_nb, coords, lwd=.2, col="blue", cex = .5)
}
