% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{st_connectivity_coords}
\alias{st_connectivity_coords}
\title{Extract coordinates of an sf polygon object for plotting a connectivity graph}
\usage{
st_connectivity_coords(x)
}
\arguments{
\item{x}{sf object}
}
\description{
Extract coordinates of an sf polygon object for plotting a connectivity graph
}
\examples{
rook_coords <- st_connectivity_coords(ncovr)
plot(ncovr_rook_nb, rook_coords, lwd=.2, col="blue", cex = .5)
}
