% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{st_queen}
\alias{st_queen}
\title{Create queen contiguity weights from sf objects}
\usage{
st_queen(a, b = a)
}
\arguments{
\item{a}{sf object}

\item{b}{(optional) sf object}
}
\value{
sgbp object representing neighbors from queen contiguity
}
\description{
Create queen contiguity weights from sf objects
}
\examples{
ncovr_queen <- st_queen(ncovr)
str(ncovr_queen)

}
