% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{st_rook}
\alias{st_rook}
\title{Create rook contiguity weights from sf objects}
\usage{
st_rook(a, b = a)
}
\arguments{
\item{a}{sf object}

\item{b}{(optional) sf object}
}
\description{
Create rook contiguity weights from sf objects
}
\examples{
ncovr_rook <- st_rook(ncovr)
str(ncovr_rook)

}
