% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{st_as_nb}
\alias{st_as_nb}
\title{Convert sparse geometric binary objects to nb objects}
\usage{
st_as_nb(sgbp)
}
\arguments{
\item{sgbp}{sparse geometric binary object, returned by st_rook or st_queen}
}
\value{
nb object
}
\description{
Convert sparse geometric binary objects to nb objects
}
\examples{
ncovr_rook <- st_rook(ncovr)
rook_nb <- st_as_nb(ncovr_rook)

str(rook_nb)
summary(rook_nb)
class(rook_nb)
}
